BUGFIX: Patch to make sasl_interactive_bind work on Python versions < 2.3.
Bug discovered/reported by David Hawes.

The problem is with the format arguments we give to PyArg_ParseTuple:

In Python 2.3+, a "I" format argument indicates that we're either converting
the Python object into a long or an unsigned int. In versions prior to that, it
will always convert to a long. Since the sasl_flags variable is an unsigned
int, we need to use the "I" flag if we're running Python 2.3+ and a "i"
otherwise.

See the email exchange:
http://sourceforge.net/mailarchive/forum.php?thread_id=6720713&forum_id=4346

deepak giridharagopal