Factor out LDAPberval_to_object()
Also commit other berval converters which are not yet called by anything.

In places where bervals are returned as objects, PyString_FromStringAndSize()
was being used. In some rare cases, a NULL berval pointer (not NULL data)
is returned as None (not a string). This factors that out to the berval.c
compilation unit.